+++ src/logger.js
@@ -7,7 +7,7 @@
  *
  * @param {string|function} [name='logger'] - The name prefix. Prepended to all output of the created logger. If it is a function, the function should return a string.
  * @param {object} [config] - An optional configuration object.
- * @param {object|array} [config.logger = window.console] - The logger object that will be used. Can also be an array of logger objects.
+ * @param {object|array} [config.backend = window.console] - The logger object that will be used. Can also be an array of logger objects.
  * @param {array} [config.functions = ['log', 'info', 'warn', 'error', 'trace', 'table', 'debug']] - An array of function names supported by the `logger` object..
  * @param {string} [config.accessor='logger'] - The name by which the logger object can be accessed on the decorated instance.
  * @param {function} [config.prefixer = (name) => `[${name}]`] - A function that creates the prefix. Receives the `name` string and should return a string.
@@ -25,28 +25,31 @@
  * }
  */
 export default function logger(name = 'logger', {
-    logger = console,
-    functions = ['log', 'info', 'warn', 'error', 'trace', 'table', 'debug'],
+    backend = console,
+    functions = ['log', 'info', 'warn', 'error', 'trace', 'table', 'debug', 'dir'],
     accessor = 'logger',
     prefixer = (name) => `[${name}]`,
     level = 'debug'
 } = {}) {
     return function decorateClass(target) {
+        if (!target) {
+            target = Object.create(null);
+            accessor = 'this';
+        }
         const customLogger = functions.reduce((result, fn) => {
-            return Object.assign(result, {
-                [fn]: (...args) => {
-                    const fnLevel = functions.indexOf(fn);
-                    const allowedLevel = typeof level === 'number' ? level : functions.indexOf(level);
-                    if (!allowedLevel || fnLevel > allowedLevel) {
-                        return;
-                    }
-                    const prefix = prefixer(typeof name === 'function' ? name({target, args}) : name);
-                    const logArgs = [prefix, ...args];
-                    const loggers = Array.isArray(logger) ? logger : [logger];
-                    loggers.forEach(logger => logger[fn].apply(logger, logArgs));
-                    return true;
+            result[fn] = (function(...args) {
+                const fnLevel = functions.indexOf(fn);
+                const allowedLevel = typeof level === 'number' ? level : functions.indexOf(level);
+                if (!allowedLevel || fnLevel > allowedLevel) {
+                    return;
                 }
-            });
+                const prefix = prefixer(typeof name === 'function' ? name({target, args}) : name);
+                const logArgs = [prefix, ...args];
+                const backends = Array.isArray(backend) ? backend : [backend];
+                backends.forEach(backend => backend[fn].apply(backend, logArgs));
+                return true;
+            }).bind(result);
+            return result;
         }, {});
         if (accessor === 'this') {
             Object.assign((target.prototype || target), customLogger);